+++ CHANGELOG.md
@@ -10,6 +10,12 @@
 
 ### Bug fixes
 
+### Contributors
+
+# Release 0.8.1
+
+### Bug fixes
+
 * Fixed a bug where gradient computations always returned 0 when
   loading a parametrized Qiskit circuit as a PennyLane template.
   [(#71)](https://github.com/XanaduAI/pennylane-qiskit/pull/71)